further enhance AWS parity, add snapshot tests, restructure codebase
