+++ airspeed/api.py
@@ -1,30 +1,13 @@
-# encoding: utf-8
-
 import os
 
-try:
-    from airspeed import CachingFileLoader
-except ImportError:
-    raise ImportError("You must install the airspeed package.")
-
 from cachetools import LRUCache
+
+from airspeed.operators import CachingFileLoader
 
 __all__ = ["Airspeed"]
 
 
 class Airspeed(object):
-
-    """The airspeed templating engine.
-
-    Sample usage:
-
-        >>> from cti.core import Engines
-        >>> render = Engines()
-        >>> render('airspeed:../tests/test.txt', dict())
-        ('text/plain', 'Bingo!')
-
-    """
-
     def __init__(self, cache=10, **kw):
         self.loaders = LRUCache(maxsize=cache)
